requirements.txt
fastapi
uvicorn
streamlit
pypdf
sentence-transformers
faiss-cpu
numpy
requests
ollama
pydantic